Exchange services, Developer controller, Provide an automated method to update products, Refactore price table value type

+++ app/Models/Price.php
@@ -15,7 +15,7 @@ class Price extends Model
     ];
 
     protected $casts = [
-        'value' => 'int',
+        'value' => 'double',
         'is_discount' => 'boolean',
     ];
